# Optimize posix `epicsThreadOnce()`